+++ Assets/Scenes/Test/PlanetBackground.unity
@@ -134,12 +134,12 @@ GameObject:
   - component: {fileID: 337582357}
   - component: {fileID: 337582356}
   m_Layer: 0
-  m_Name: BackgorundVisualEffects
+  m_Name: PlanetBackground
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!114 &337582356
 MonoBehaviour:
   m_ObjectHideFlags: 0
